+++ src/modules.mjs
@@ -120,6 +120,7 @@ function calculateLibraries() {
         'libpipefs.js', // ok to include it by default since it's only used if the syscall is used
         'libsockfs.js', // ok to include it by default since it's only used if the syscall is used
         'libepoll.js', // ok to include it by default since it's only used if the syscall is used
+        'libtimerfd.js', // ok to include it by default since it's only used if the syscall is used
       );
 
       if (NODERAWSOCKETS) {